+++ src/traits/to_u16.rs
@@ -101,7 +101,7 @@ mod tests {
 
             for &value in VALUES {
                 let expected = value;
-                let instance = CustomType { value: value as u16 };
+                let instance = CustomType { value: value };
                 let actual = instance.to_u16();
 
                 assert_eq!(expected, actual);
